% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print the  name of an animal and his sound based on the user input}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A \strong{character} string specifying the animal name}

\item{sound}{A \strong{character} string specifying the name of the sound}
}
\value{
A \emph{character} string containing a sentence with the animal name and the sound
}
\description{
Print a sentence saying what sound an animal makes from user inputs.
}
\examples{

animal_sounds("cow","moo")

}
