% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsound-package.R
\docType{package}
\name{animalsound-package}
\alias{animalsound}
\alias{animalsound-package}
\title{animalsound: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Shenggang/animalsound}
  \item Report bugs at \url{https://github.com/Shenggang/animalsound/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
