% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsounds.R
\name{animalsounds}
\alias{animalsounds}
\title{Title Test package Animal sounds}
\usage{
animalsounds(animal = NULL, sound = NULL)
}
\arguments{
\item{animal}{A character variable giving names of animals}

\item{sound}{A character variable giving corresponding sounds animal make}

\item{country}{A character variable giving information on the country}
}
\value{
A vector with the values pasted together
}
\description{
Title Test package Animal sounds
}
\examples{

animalsounds("dog","ouah ouah")

}
