% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{farm_animals dataset}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{animals}{Name of the animal}
\item{sounds}{Sound of the animal}
}
}
\source{
Hand-coded common sense.
}
\usage{
farm_animals
}
\description{
A hand-coded list of farm animals and their sounds.
}
\keyword{datasets}
