% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{animalsound package}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A non-numerical string of length greater than 1}

\item{sound}{A non-numerical string of length greater than 1}
}
\value{
A string composed using the inputs
}
\description{
animalsound function, part of R Workshop example package
}
\examples{
animal_sounds('dog', 'woolf')
}
