% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{animalsound package}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{A non-numerical string of length greater than 1}

\item{sound}{A non-numerical string of length greater than 1}
}
\value{
A string composed using the inputs
}
\description{
animalsound function, part of R Workshop example package.
This is an example of text in  \emph{\strong{markdown}}.
}
\details{
\subsection{Detail Section Header}{

Some other markdowns we can use
\enumerate{
\item Lists
\item Enumerated Lists
\item Tables
\item Code snippts
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{animal_sounds("abc", "cdf")
}\if{html}{\out{</div>}}

The function uses \code{\link[=paste0]{paste0()}} to form the output string.
}
}
\examples{
animal_sounds('dog', 'woolf')
}
