% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsound-package.R
\docType{package}
\name{animalsound-package}
\alias{animalsound}
\alias{animalsound-package}
\title{animalsound: Animal Sound (Workshop Example Package)}
\description{
A package built while going through the development practices in the course material.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Shenggang/animalsound}
  \item Report bugs at \url{https://github.com/Shenggang/animalsound/issues}
}

}
\author{
\strong{Maintainer}: Shenggang Hu \email{hushenggang1996@gmail.com}

}
\keyword{internal}
